requirements.txt
absl-py==2.4.0
altair==6.0.0
astunparse==1.6.3
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.62.1
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.14.0
kiwisolver==1.5.0
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
narwhals==2.18.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
pillow==12.2.0
protobuf==4.25.9
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
referencing==0.37.0
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
tenacity==9.1.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
